organize files to improve file structure. Remove excess code from Homepage
